let timer service report perf telemetry, https://github.com/microsoft/vscode/issues/101850
